Refactor shared API functions out into `src/headless/shared/api`
